Refactor tests to improve message handling consistency
